Time: 4 ms (13.65%), Space: 42.3 MB (48.45%) - LeetHub

+++ 1544-make-the-string-great/1544-make-the-string-great.java
@@ -3,12 +3,12 @@ class Solution {
         if (s.length() < 2) {
             return s;
         }
-        
+
         Stack<Character> stack = new Stack<>();
 
         for(var ch : s.toCharArray()){
             if(!stack.empty()){
-                if(Math.abs((int)ch - ((int)stack.peek())) == 32){
+                if(Math.abs(ch - (stack.peek())) == 32){
                     stack.pop();
                     continue;
                 }
